+++ apps/dashboard/app/(app)/settings/billing/stripe/portal/page.tsx
@@ -7,16 +7,14 @@ import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import Stripe from "stripe";
 
-export const revalidate = 0;
+export const dynamic = "force-dynamic";
 
 export default async function StripeRedirect() {
   const { orgId } = await getAuth();
-  if (!orgId) {
-    return redirect("/auth/sign-in");
-  }
 
   const ws = await db.query.workspaces.findFirst({
-    where: (table, { and, eq, isNull }) => and(eq(table.orgId, orgId), isNull(table.deletedAtM)),
+    where: (table, { and, eq, isNull }) =>
+      and(eq(table.orgId, orgId), isNull(table.deletedAtM)),
   });
   if (!ws) {
     return redirect("/new");
@@ -27,7 +25,8 @@ export default async function StripeRedirect() {
       <Empty>
         <Empty.Title>Stripe is not configured</Empty.Title>
         <Empty.Description>
-          If you are selfhosting Unkey, you need to configure Stripe in your environment variables.
+          If you are selfhosting Unkey, you need to configure Stripe in your
+          environment variables.
         </Empty.Description>
       </Empty>
     );
